campus_backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
redis==5.0.3
